Rename dnd to drag-n-drop for better readability + Move AppWidget to src root, out of drag-n-drop
