+++ src/drag-n-drop/ComponentLibraryWidget.tsx
