+++ src/drag-n-drop/LibraryItemWidget.tsx
